+++ CHANGELOG.md
@@ -23,6 +23,11 @@
   (PR[#3697](https://github.com/scality/metalk8s/pull/3697))
 
 ## Release 2.11.3 (in development)
+### Bug fixes
+
+- Fix a bug during the upgrade that makes the workload plane Ingress controller
+  ignore the Ingress object that does not have the class explicitly set
+  (PR[#3704](https://github.com/scality/metalk8s/pull/3704))
 
 ## Release 2.11.2
 ### Bug fixes